extractor/requirements.txt
streamlit
firebase-admin
requests
Pillow
